requirements.txt
pandas
hopsworks
python-dotenv
parsedatetime
scikit-learn
ta
numpy
fire
optuna
xgboost
lightgbm
comet_ml